requirements.txt
openai>=1.0.0
PyQt6>=6.6.0
mss>=9.0.0
keyboard>=0.13.5
python-dotenv>=1.0.0
Pillow>=10.0.0
sounddevice>=0.5.0
numpy>=1.24.0
